requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-multipart
jinja2
openai
faster-whisper
python-dotenv
